rename subpackage mesh to discovery
